@@ -13,52 +13,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular/flex-layout": "7.0.0-beta.23",
-    "@angular/forms": "7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
-    "@auth0/angular-jwt": "^4.2.0",
-    "@ngx-translate/core": "^11.0.1",
-    "@ngx-translate/http-loader": "^4.0.0",
+    "@angular/animations": "9.1.12",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "9.1.12",
+    "@angular/compiler": "9.1.12",
+    "@angular/core": "9.1.12",
+    "@angular/elements": "^9.1.12",
+    "@angular/forms": "9.1.12",
+    "@angular/platform-browser": "9.1.12",
+    "@angular/platform-browser-dynamic": "9.1.12",
+    "@angular/router": "9.1.12",
+    "@auth0/angular-jwt": "^5.0.1",
+    "@ngx-translate/core": "^12.1.2",
+    "@ngx-translate/http-loader": "^5.0.0",
     "@types/zxcvbn": "^4.4.0",
     "ajv": "^6.12.4",
-    "angular-formio": "^3.27.3",
-    "angular-password-strength-meter": "0.0.4",
+    "angular-formio": "^4.9.6",
+    "angular-password-strength-meter": "2.0.0",
     "bootstrap": "^3.4.1",
-    "formiojs": "^3.29.18",
+    "document-register-element": "^1.7.2",
+    "formiojs": "^4.11.3",
     "jasmine": "^2.99.0",
     "jquery": "^3.5.1",
     "lodash": "^4.17.20",
     "ng-event-source": "^1.0.14",
-    "ng-recaptcha": "^4.3.0",
+    "ng-recaptcha": "^5.0.0",
     "ng-terminal": "^2.4.0",
-    "ng2-tooltip-directive": "2.8.17",
+    "ng2-tooltip-directive": "2.9.20",
     "ngx-pagination": "^3.3.1",
     "ngx-webstorage-service": "^4.1.0",
     "primeicons": "^1.0.0",
-    "primeng": "^7.1.3",
+    "primeng": "^9.1.3",
     "rxjs": "^6.6.3",
-    "tslib": "^2.0.1",
-    "zone.js": "^0.8.29",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.10",
-    "@angular/cli": "^7.3.10",
-    "@angular/compiler-cli": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.901.12",
+    "@angular/cli": "^9.1.12",
+    "@angular/compiler-cli": "^9.1.12",
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
     "@types/jasmine": "^2.8.17",
-    "@types/node": "^9.6.58",
+    "@types/node": "^12.11.1",
     "babel-preset-env": "^1.7.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.1.2",
     "intl": "^1.2.5",
     "jasmine-core": "^2.99.1",
     "jasmine-spec-reporter": "4.2.1",
@@ -75,7 +75,7 @@
     "puppeteer": "^1.20.0",
     "ts-node": "^4.1.0",
     "tslint": "^5.20.1",
-    "typescript": "3.2.4",
+    "typescript": "3.8.3",
     "webpack": "^4.44.1"
   }
 }
