@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LanguageListComponent } from './languagelist/languagelist.component';
 import { LanguageDetailsComponent } from './languagedetails/languagedetails.component';
-import {InternationalizationService} from "../../../service/internationalization.service";
-import {RouterModule} from "@angular/router";
-import {TranslateModule} from "@ngx-translate/core";
-import {SharedModule} from "../../../shared";
-import {FormsModule} from "@angular/forms";
-import {InputSwitchModule} from "primeng/primeng";
+import {InternationalizationService} from '../../../service/internationalization.service';
+import {RouterModule} from '@angular/router';
+import {TranslateModule} from '@ngx-translate/core';
+import {SharedModule} from '../../../shared';
+import {FormsModule} from '@angular/forms';
+import {InputSwitchModule} from 'primeng/inputswitch';
 
 @NgModule({
   declarations: [LanguageListComponent, LanguageDetailsComponent],
