@@ -1,6 +1,6 @@
 import {Component, Input, OnInit} from '@angular/core';
 
-import {AppInstanceProgressStage, AppInstanceState} from '../../../model/index';
+import {AppInstanceProgressStage, AppInstanceState} from '../../../model';
 import {TranslateService} from '@ngx-translate/core';
 
 @Component({
@@ -30,6 +30,32 @@ export class AppInstanceProgressComponent implements OnInit {
     getTranslateTag(stateProgress): string {
         stateProgress = stateProgress.toString().toUpperCase().split(' ').join('_');
         return this.translate.instant('APP_INSTANCE.PROGRESS.' + stateProgress.toString());
+    }
+
+    // future not failed state
+    public displaySuccessStyle(stage: AppInstanceProgressStage): boolean {
+        return (stage.activeState < this.activeState && this.activeState !== AppInstanceState.FAILURE)
+            || (stage.activeState < this.previousState && this.previousState !== AppInstanceState.UNKNOWN)
+    }
+
+    // current not failed state
+    public displayPrimaryStyle(stage: AppInstanceProgressStage): boolean {
+        return stage.activeState === this.activeState && this.activeState !== AppInstanceState.FAILURE;
+    }
+
+    // failed or removed app instance
+    public displayDangerStyle(stage: AppInstanceProgressStage): boolean {
+        return this.activeState === AppInstanceState.FAILURE
+            && stage.activeState === this.previousState
+            || this.activeState === AppInstanceState.REMOVED
+    }
+
+    public displayDefaultStyle(stage: AppInstanceProgressStage): boolean {
+        return (stage.activeState > this.previousState && this.previousState !== AppInstanceState.UNKNOWN )
+            || (stage.activeState > this.activeState && this.activeState !== AppInstanceState.FAILURE)
+            || (this.activeState === AppInstanceState.UNKNOWN)
+            || (stage.activeState === AppInstanceState.REQUESTED)
+            || (!stage?.activeState);
     }
 
 }
