@@ -9,7 +9,6 @@ import {TranslateModule} from '@ngx-translate/core';
 import {AppsService, TagService} from '../../service';
 import {AppManagementListComponent} from './appmanagementlist/appmanagementlist.component';
 import {StepsModule} from 'primeng/steps';
-import {ChipsModule, FileUploadModule, MultiSelectModule} from 'primeng/primeng';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {AppCreateWizardComponent} from './appcreatewizard/appcreatewizard.component';
 import {InternationalizationService} from '../../service/internationalization.service';
@@ -22,6 +21,10 @@ import {AppVersionCreateWizardComponent} from './appversioncreatewizard/appversi
 import {AppAccessMethodEditComponent} from './app-access-method-edit-component/app-access-method-edit.component';
 import {AppStorageVolumeEditComponent} from './app-storage-volume-edit-component/app-storage-volume-edit.component';
 import {AppStaticGlobalDeployParametersEditComponent} from './app-static-global-deploy-parameters-edit/app-static-global-deploy-parameters-edit.component';
+import {MultiSelectModule} from 'primeng/multiselect';
+import {FileUploadModule} from 'primeng/fileupload';
+import {ChipsModule} from 'primeng/chips';
+
 
 export function getJsonTemplates(config: ConfigTemplateService) {
     return () => config.loadConfigTemplate();
