@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { RegistrationComponent } from './registration.component';
-import {RecaptchaFormsModule} from 'ng-recaptcha/forms';
+import {RecaptchaFormsModule} from 'ng-recaptcha';
 
 describe('RegistrationComponent', () => {
   let component: RegistrationComponent;
